# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trivelo Hotel Booking API",
        "description": "Comprehensive hotel booking system API with multi-role authentication",
        "contact": {
            "email": "support@trivelo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List rooms for a hotel",
                "description": "Get list of active rooms for a specific hotel with filtering",
                "operationId": "db2a0f9c242ba089e87d5ae1fa8e14c5",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "Hotel ID (required)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "room_type",
                        "in": "query",
                        "description": "Filter by room type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "standard",
                                "deluxe",
                                "suite",
                                "presidential"
                            ]
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum room price per night",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum room price per night",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Minimum room capacity",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rooms retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error - hotel_id required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "description": "Create a new room for a hotel (Hotel Manager only)",
                "operationId": "809158a6b73bac8545f93f5b9f8a8a8a",
                "requestBody": {
                    "description": "Room creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoomRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Only hotel managers can create rooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRoomRequest": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/RoomCreateRequest"
                    }
                ]
            },
...
            "RoomListResponse": {
                "title": "Room List Response",
                "description": "Paginated room list response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Rooms retrieved successfully"
                    },
                    "data": {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RoomSummary"
                                }
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 42
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "RoomDetailResponse": {
                "title": "Room Detail Response",
                "description": "Single room detail response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Room details retrieved successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/RoomDetail"
                    }
                },
                "type": "object"
            },
            "RoomSummary": {
                "title": "Room Summary",
                "description": "Room summary information",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_number": {
                        "type": "string",
                        "example": "101"
                    },
                    "room_type": {
                        "type": "string",
                        "example": "deluxe"
                    },
                    "price_per_night": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Error API response format",
                "properties": {
                    "success": {
                        "description": "Operation success status",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Detailed error information",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "The email field is required.",
                            "The password must be at least 8 characters."
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}